@@ -1,74 +1,79 @@
-buildscript {
-    repositories {
-        maven { url 'https://maven.fabric.io/public' }
-    }
-
-    dependencies {
-        classpath 'io.fabric.tools:gradle:1.+'
-    }
-}
 apply plugin: 'com.android.application'
-apply plugin: 'io.fabric'
-apply plugin: 'android-apt'
 
 repositories {
-    maven { url 'https://maven.fabric.io/public' }
+    google()
 }
 
-
 android {
-    compileSdkVersion 25
-    buildToolsVersion "25.0.0"
+    compileSdkVersion 30
+    buildToolsVersion '30.0.2'
 
     defaultConfig {
         applicationId "tech.salroid.filmy"
-        minSdkVersion 16
-        targetSdkVersion 25
-        versionCode 11
-        versionName "2.0.1"
+        minSdkVersion 21
+        targetSdkVersion 30
+        versionCode 17
+        versionName "2.1.3"
         multiDexEnabled true
-
+        resConfigs "en", "de", "es-rES", "hi-rIN", "nl-rNL"
     }
     buildTypes {
         release {
             minifyEnabled true
-            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
-            buildConfigField("String", "API_KEY", "\"" + getApiKey() + "\"")
+            shrinkResources true
+            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
+            buildConfigField("String", "TMDB_API_KEY", "\"" + getTMDBApiKey() + "\"")
+            buildConfigField("String", "OMDB_API_KEY", "\"" + getOMDBApiKey() + "\"")
         }
-        debug{
-            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
-            buildConfigField("String", "API_KEY", "\"" + getApiKey() +"\"")
+        debug {
+            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
+            buildConfigField("String", "TMDB_API_KEY", "\"" + getTMDBApiKey() + "\"")
+            buildConfigField("String", "OMDB_API_KEY", "\"" + getOMDBApiKey() + "\"")
             minifyEnabled true
         }
     }
+
+    compileOptions {
+        sourceCompatibility 1.8
+        targetCompatibility 1.8
+    }
 }
 
-def getApiKey(){
-    def Properties props = new Properties()
+static def getTMDBApiKey() {
+    Properties props = new Properties()
     props.load(new FileInputStream(new File('secrets.properties')))
     return props['TMDB_API_KEY']
 }
 
+static def getOMDBApiKey() {
+    Properties props = new Properties()
+    props.load(new FileInputStream(new File('secrets.properties')))
+    return props['OMDB_API_KEY']
+}
+
+
 dependencies {
-    compile fileTree(dir: 'libs', include: ['*.jar'])
-    compile 'com.android.support:appcompat-v7:25.3.1'
-    compile 'com.android.support:design:25.3.1'
-    compile 'com.android.support:customtabs:25.3.1'
-    compile 'com.github.bumptech.glide:glide:3.7.0'
-    compile 'com.android.volley:volley:1.0.0'
-    compile 'com.mikhaellopez:circularimageview:3.0.2'
-    compile 'com.android.support:palette-v7:25.3.1'
-    compile 'com.miguelcatalan:materialsearchview:1.4.0'
-    compile 'com.android.support:cardview-v7:25.3.1'
-    testCompile 'junit:junit:4.12'
-    compile 'me.tatarka.support:jobscheduler:0.1.1'
-    compile 'com.github.xiprox.errorview:library:2.+'
-    compile 'com.github.paolorotolo:appintro:4.0.0'
-    compile('com.crashlytics.sdk.android:crashlytics:2.6.7@aar') {
-        transitive = true;
-    }
-    compile 'com.jakewharton:butterknife:8.4.0'
-    apt 'com.jakewharton:butterknife-compiler:8.4.0'
-    compile 'com.google.firebase:firebase-appindexing:10.0.0'
-    compile files('libs/YouTubeAndroidPlayerApi.jar')
+    implementation fileTree(dir: 'libs', include: ['*.jar'])
+
+    implementation 'com.google.android.material:material:1.3.0-alpha03'
+    implementation 'androidx.palette:palette:1.0.0'
+    implementation 'androidx.browser:browser:1.3.0'
+    implementation "androidx.work:work-runtime:2.4.0"
+
+    implementation 'com.github.bumptech.glide:glide:4.11.0'
+    implementation 'com.mcxiaoke.volley:library:1.0.19'
+    implementation 'com.mikhaellopez:circularimageview:3.0.2'
+    implementation 'com.miguelcatalan:materialsearchview:1.4.0'
+    implementation 'tr.xip.errorview:library:3.0.0'
+    implementation 'com.github.paolorotolo:appintro:4.0.0'
+
+
+    implementation 'com.jakewharton:butterknife:10.1.0'
+    annotationProcessor 'com.jakewharton:butterknife-compiler:10.1.0'
+
+    implementation 'com.google.firebase:firebase-appindexing:20.0.0'
+    implementation files('libs/YouTubeAndroidPlayerApi.jar')
+    debugImplementation 'com.amitshekhar.android:debug-db:1.0.6'
+    testImplementation 'junit:junit:4.13.1'
+    androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
 }
