@@ -2,9 +2,12 @@ package tech.salroid.filmy;
 
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
+
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+
 import com.github.paolorotolo.appintro.AppIntro2;
+
 import tech.salroid.filmy.fragment.intro_fragments.IntroFragmentA;
 import tech.salroid.filmy.fragment.intro_fragments.IntroFragmentB;
 import tech.salroid.filmy.fragment.intro_fragments.IntroFragmentC;
