@@ -4,10 +4,12 @@ import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.Bundle;
 import android.preference.PreferenceManager;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.Toolbar;
+import android.view.MenuItem;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 
 import java.util.List;
 
@@ -37,7 +39,8 @@ import tech.salroid.filmy.parser.CharacterDetailActivityParseWork;
 
 public class FullMovieActivity extends AppCompatActivity implements CharacterDetailsActivityAdapter.ClickListener {
 
-    @BindView(R.id.toolbar) Toolbar toolbar;
+    @BindView(R.id.toolbar)
+    Toolbar toolbar;
     @BindView(R.id.full_movie_recycler)
     RecyclerView full_movie_recycler;
     private String movie_result;
@@ -60,15 +63,18 @@ public class FullMovieActivity extends AppCompatActivity implements CharacterDet
         setSupportActionBar(toolbar);
 
 
+
         full_movie_recycler.setLayoutManager(new LinearLayoutManager(FullMovieActivity.this));
 
 
         Intent intent = getIntent();
         if (intent != null) {
             movie_result = intent.getStringExtra("cast_json");
-            if (getSupportActionBar() != null)
+            if (getSupportActionBar() != null) {
+                getSupportActionBar().setDisplayHomeAsUpEnabled(true);
                 getSupportActionBar().setTitle(intent.getStringExtra("toolbar_title"));
-        }
+            }
+            }
 
 
         CharacterDetailActivityParseWork par = new CharacterDetailActivityParseWork(this, movie_result);
@@ -84,6 +90,7 @@ public class FullMovieActivity extends AppCompatActivity implements CharacterDet
     public void itemClicked(CharacterDetailsData setterGetterChar, int position) {
         Intent intent = new Intent(this, MovieDetailsActivity.class);
         intent.putExtra("id", setterGetterChar.getChar_id());
+        intent.putExtra("title",setterGetterChar.getChar_movie());
         intent.putExtra("network_applicable", true);
         intent.putExtra("activity", false);
         startActivity(intent);
@@ -99,5 +106,15 @@ public class FullMovieActivity extends AppCompatActivity implements CharacterDet
         boolean nightModeNew = sp.getBoolean("dark", false);
         if (nightMode!=nightModeNew)
             recreate();
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+
+
+        if (item.getItemId() == android.R.id.home)
+            finish();
+
+        return super.onOptionsItemSelected(item);
     }
 }
