@@ -7,7 +7,9 @@ import android.content.UriMatcher;
 import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
 import android.net.Uri;
-import android.support.annotation.Nullable;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 import tech.salroid.filmy.database.FilmContract;
 import tech.salroid.filmy.database.FilmDbHelper;
@@ -41,8 +43,10 @@ public class FilmProvider extends ContentProvider {
     static final int CAST = 300;
     static final int SAVE = 200;
     static final int SAVE_DETAILS_WITH_MOVIE_ID = 201;
+
     // The URI Matcher used by this content provider.
     private static final UriMatcher sUriMatcher = buildUriMatcher();
+
     private static final String sTrendingMovieIdSelection =
             FilmContract.MoviesEntry.TABLE_NAME +
                     "." + FilmContract.MoviesEntry.MOVIE_ID + " = ? ";
@@ -329,7 +333,7 @@ public class FilmProvider extends ContentProvider {
 
     @Nullable
     @Override
-    public Uri insert(Uri uri, ContentValues contentValues) {
+    public Uri insert(@NonNull Uri uri, ContentValues contentValues) {
 
         final SQLiteDatabase db = mOpenHelper.getWritableDatabase();
         final int match = sUriMatcher.match(uri);
