@@ -5,22 +5,22 @@ import android.content.SharedPreferences;
 import android.os.Build;
 import android.os.Bundle;
 import android.preference.PreferenceManager;
-import android.support.v4.app.ActivityOptionsCompat;
-import android.support.v4.util.Pair;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.Toolbar;
+import android.view.MenuItem;
 import android.view.View;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
+import androidx.core.app.ActivityOptionsCompat;
+import androidx.core.util.Pair;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 
 import java.util.List;
 
 import butterknife.BindView;
 import butterknife.ButterKnife;
 import tech.salroid.filmy.R;
-import tech.salroid.filmy.custom_adapter.CastAdapter;
 import tech.salroid.filmy.custom_adapter.CrewAdapter;
-import tech.salroid.filmy.data_classes.CastDetailsData;
 import tech.salroid.filmy.data_classes.CrewDetailsData;
 import tech.salroid.filmy.parser.MovieDetailsActivityParseWork;
 
@@ -43,8 +43,10 @@ import tech.salroid.filmy.parser.MovieDetailsActivityParseWork;
 
 public class FullCrewActivity extends AppCompatActivity implements CrewAdapter.ClickListener {
 
-    @BindView(R.id.toolbar) Toolbar toolbar;
-    @BindView(R.id.full_cast_recycler) RecyclerView full_crew_recycler;
+    @BindView(R.id.toolbar)
+    Toolbar toolbar;
+    @BindView(R.id.full_cast_recycler)
+    RecyclerView full_crew_recycler;
 
 
     private String crew_result;
@@ -73,8 +75,10 @@ public class FullCrewActivity extends AppCompatActivity implements CrewAdapter.C
         Intent intent = getIntent();
         if (intent != null) {
             crew_result = intent.getStringExtra("crew_json");
-            if(getSupportActionBar()!=null)
-            getSupportActionBar().setTitle(intent.getStringExtra("toolbar_title"));
+            if (getSupportActionBar() != null) {
+                getSupportActionBar().setDisplayHomeAsUpEnabled(true);
+                getSupportActionBar().setTitle(intent.getStringExtra("toolbar_title"));
+            }
         }
 
 
@@ -90,7 +94,7 @@ public class FullCrewActivity extends AppCompatActivity implements CrewAdapter.C
     public void itemClicked(CrewDetailsData setterGetter, int position, View view) {
 
         Intent intent = new Intent(this, CharacterDetailsActivity.class);
-        intent.putExtra("id", setterGetter.getCrew_id());
+        intent.putExtra("id", setterGetter.getCrewId());
 
         if (Build.VERSION.SDK_INT > Build.VERSION_CODES.KITKAT) {
 
@@ -114,5 +118,15 @@ public class FullCrewActivity extends AppCompatActivity implements CrewAdapter.C
         boolean nightModeNew = sp.getBoolean("dark", false);
         if (nightMode!=nightModeNew)
             recreate();
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+
+
+        if (item.getItemId() == android.R.id.home)
+            finish();
+
+        return super.onOptionsItemSelected(item);
     }
 }
