@@ -1,16 +1,15 @@
 package tech.salroid.filmy.activities;
 
 import android.content.Context;
-
-import com.android.volley.Request;
 import com.android.volley.RequestQueue;
 import com.android.volley.Response;
 import com.android.volley.VolleyError;
 import com.android.volley.toolbox.JsonObjectRequest;
-
+import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
 
+import tech.salroid.filmy.BuildConfig;
 import tech.salroid.filmy.network_stuff.VolleySingleton;
 
 /*
@@ -30,35 +29,54 @@ import tech.salroid.filmy.network_stuff.VolleySingleton;
  * limitations under the License.
  */
 
-public class Rating {
+class Rating {
 
-    public static String imdb_rating = "0",tomatometer_rating="0",audience_rating="0",metascore_rating="0",image=null;
+    private static String imdb_rating = "0", tomatometer_rating = "0", audience_rating = "0", metascore_rating = "0", image = null, rottenTomatoPage = null;
+    private static String OMDB_API_KEY = BuildConfig.OMDB_API_KEY;
 
-    public static void getRating(final Context context, String movie_id_final) {
+    static void getRating(final Context context, String movie_id_final) {
 
         VolleySingleton volleySingleton = VolleySingleton.getInstance();
         RequestQueue requestQueue = volleySingleton.getRequestQueue();
 
 
-        String BASE_RATING_URL = "http://www.omdbapi.com/?i=" + movie_id_final + "&tomatoes=true&r=json";
+        String BASE_RATING_URL = "http://www.omdbapi.com/?i=" + movie_id_final + "&apikey=" + OMDB_API_KEY + "&tomatoes=true&r=json";
 
-        JsonObjectRequest jsonObjectRequestForMovieDetails = new JsonObjectRequest(Request.Method.GET, BASE_RATING_URL, null,
+        JsonObjectRequest jsonObjectRequestForMovieDetails = new JsonObjectRequest(BASE_RATING_URL, null,
                 new Response.Listener<JSONObject>() {
                     @Override
                     public void onResponse(JSONObject response) {
-                        try {
-                            imdb_rating = response.getString("imdbRating");
-                            tomatometer_rating=response.getString("tomatoRating");
-                            audience_rating=response.getString("tomatoUserRating");
-                            metascore_rating=response.getString("Metascore");
-                            image=response.getString("tomatoImage");
 
-                            setRatingCallback(context,imdb_rating,tomatometer_rating,audience_rating,metascore_rating,image);
+                        try {
+
+                            Boolean responseBool = response.getBoolean("Response");
+
+                            if (responseBool){
+
+                                imdb_rating = response.getString("imdbRating");
+                                tomatometer_rating = response.getString("tomatoRating");
+                                audience_rating = response.getString("tomatoUserRating");
+                                metascore_rating = response.getString("Metascore");
+                                image = response.getString("tomatoImage");
+                                rottenTomatoPage = response.getString("tomatoURL");
+
+
+                                //Above tomatometer does not work this does
+                                JSONArray jsonArray = response.getJSONArray("Ratings");
+                                for (int i = 0; i < jsonArray.length(); i++) {
+                                    if (jsonArray.getJSONObject(i).getString("Source").equals("Rotten Tomatoes")) {
+                                        tomatometer_rating = jsonArray.getJSONObject(i).getString("Value");
+                                    }
+                                }
+
+                                setRatingCallback(context, imdb_rating, tomatometer_rating, audience_rating, metascore_rating, rottenTomatoPage);
+
+                            }else
+                                setRatingFailCallback(context);
 
                         } catch (JSONException e) {
                             e.printStackTrace();
-                            setRatingCallback(context,imdb_rating,tomatometer_rating,audience_rating,metascore_rating,image);
-
+                            setRatingFailCallback(context);
                         }
 
                     }
@@ -69,7 +87,7 @@ public class Rating {
             public void onErrorResponse(VolleyError error) {
 
                 // Log.e("webi", "Volley Error: " + error.getCause());
-
+                setRatingFailCallback(context);
             }
         }
         );
@@ -78,11 +96,13 @@ public class Rating {
 
     }
 
+    private static void setRatingFailCallback(Context context) {
+        ((MovieDetailsActivity) context).setRatingGone();
+    }
 
-    private static void setRatingCallback(Context context,String imdb_rating,String tomatometer_rating,String audience_rating,String metascore_rating,String image) {
 
-        ((MovieDetailsActivity)context).setRating(imdb_rating,tomatometer_rating,audience_rating,metascore_rating,image);
-
+    private static void setRatingCallback(Context context, String imdb_rating, String tomatometer_rating, String audience_rating, String metascore_rating, String rottenTomatoPage) {
+        ((MovieDetailsActivity) context).setRating(imdb_rating, tomatometer_rating, audience_rating, metascore_rating, rottenTomatoPage );
     }
 
 }
