@@ -10,10 +10,12 @@ import android.preference.Preference;
 import android.preference.PreferenceFragment;
 import android.preference.PreferenceManager;
 import android.preference.SwitchPreference;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
 import android.view.MenuItem;
 import android.widget.TextView;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
+import androidx.core.content.res.ResourcesCompat;
 
 import butterknife.BindView;
 import butterknife.ButterKnife;
@@ -68,7 +70,7 @@ public class SettingsActivity extends AppCompatActivity {
         }
 
 
-        Typeface typeface = Typeface.createFromAsset(getAssets(), "fonts/canaro_extra_bold.otf");
+        Typeface typeface =  ResourcesCompat.getFont(this,R.font.rubik);
         logo.setTypeface(typeface);
 
 
@@ -119,9 +121,9 @@ public class SettingsActivity extends AppCompatActivity {
                     SwitchPreference switchPreference = (SwitchPreference) preference;
 
                     if (!switchPreference.isChecked()) {
-                        quality = "w1000";
+                        quality = "original";
                     } else {
-                        quality = "w780";
+                        quality = "w1280";
                     }
 
 
@@ -144,8 +146,8 @@ public class SettingsActivity extends AppCompatActivity {
                 }
             });
 
-            Preference liscence = (Preference) findPreference("license");
-            liscence.setOnPreferenceClickListener(new Preference.OnPreferenceClickListener() {
+            Preference license = findPreference("license");
+            license.setOnPreferenceClickListener(new Preference.OnPreferenceClickListener() {
                 @Override
                 public boolean onPreferenceClick(Preference preference) {
                     startActivity(new Intent(getActivity(), License.class));
@@ -154,13 +156,13 @@ public class SettingsActivity extends AppCompatActivity {
             });
 
 
-            Preference share = (Preference) findPreference("Share");
+            Preference share = findPreference("Share");
             share.setOnPreferenceClickListener(new Preference.OnPreferenceClickListener() {
                 @Override
                 public boolean onPreferenceClick(Preference preference) {
 
                     String app_share_details = getResources().getString(R.string.app_share_link);
-                    if (!(app_share_details.equals(null))) {
+                    if (app_share_details != null) {
                         Intent myIntent = new Intent(Intent.ACTION_SEND);
                         myIntent.setType("text/plain");
                         myIntent.putExtra(Intent.EXTRA_TEXT, "Check out this awesome movie app.\n" + "*filmy*\n" + app_share_details);
@@ -171,7 +173,7 @@ public class SettingsActivity extends AppCompatActivity {
             });
 
 
-            Preference about = (Preference) findPreference("About");
+            Preference about = findPreference("About");
             about.setOnPreferenceClickListener(new Preference.OnPreferenceClickListener() {
                 @Override
                 public boolean onPreferenceClick(Preference preference) {
