@@ -1,16 +1,20 @@
 package tech.salroid.filmy.custom_adapter;
 
 import android.content.Context;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
+
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.RecyclerView;
+
 import com.bumptech.glide.Glide;
 import com.bumptech.glide.load.engine.DiskCacheStrategy;
 import com.mikhaellopez.circularimageview.CircularImageView;
-import java.util.ArrayList;
+
 import java.util.List;
+
 import butterknife.BindView;
 import butterknife.ButterKnife;
 import tech.salroid.filmy.R;
@@ -37,7 +41,7 @@ public class CharacterDetailsActivityAdapter extends RecyclerView.Adapter<Charac
     private final LayoutInflater inflater;
     private final Boolean ret_size;
     private Context con;
-    private List<CharacterDetailsData> ch = new ArrayList<>();
+    private List<CharacterDetailsData> ch;
     private ClickListener clickListener;
 
     public CharacterDetailsActivityAdapter(Context context, List<CharacterDetailsData> ch, Boolean size) {
@@ -45,20 +49,18 @@ public class CharacterDetailsActivityAdapter extends RecyclerView.Adapter<Charac
         con = context;
         this.ch = ch;
         this.ret_size = size;
-
-
     }
 
 
     @Override
-    public Fo onCreateViewHolder(ViewGroup parent, int viewType) {
+    public Fo onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
         View view = inflater.inflate(R.layout.char_custom_row, parent, false);
 
         return new Fo(view);
     }
 
     @Override
-    public void onBindViewHolder(Fo holder, int position) {
+    public void onBindViewHolder(@NonNull Fo holder, int position) {
 
         String m_name, m_profile, m_desc;
         // String m_id;
@@ -112,7 +114,6 @@ public class CharacterDetailsActivityAdapter extends RecyclerView.Adapter<Charac
         Fo(View itemView) {
             super(itemView);
             ButterKnife.bind(this, itemView);
-
 
             itemView.setOnClickListener(new View.OnClickListener() {
                 @Override
