@@ -3,13 +3,14 @@ package tech.salroid.filmy.custom_adapter;
 
 import android.content.Context;
 import android.database.Cursor;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.FrameLayout;
 import android.widget.ImageView;
 import android.widget.TextView;
+
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.bumptech.glide.Glide;
 import com.bumptech.glide.load.engine.DiskCacheStrategy;
@@ -77,7 +78,6 @@ public class SavedMoviesAdapter extends RecyclerView.Adapter<SavedMoviesAdapter.
         movie_poster = dataCursor.getString(poster_index);
         imdb_id = dataCursor.getString(id_index);
         movie_year = dataCursor.getInt(year_index);
-
 
         holder.title.setText(movie_title);
 
@@ -153,7 +153,8 @@ public class SavedMoviesAdapter extends RecyclerView.Adapter<SavedMoviesAdapter.
                     int movie_title_index = dataCursor.getColumnIndex(FilmContract.SaveEntry.SAVE_TITLE);
 
                     if (clickListener != null) {
-                        clickListener.itemClicked(dataCursor.getString(movie_id_index), dataCursor.getString(movie_title_index));
+                        clickListener.itemClicked(dataCursor.getString(movie_id_index),
+                                dataCursor.getString(movie_title_index));
                     }
 
                 }
