@@ -3,25 +3,23 @@ package tech.salroid.filmy.fragment;
 import android.content.Intent;
 import android.os.Build;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v4.app.ActivityOptionsCompat;
-import android.support.v4.app.Fragment;
-import android.support.v4.util.Pair;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.RelativeLayout;
 import android.widget.TextView;
 
-import com.android.volley.Request;
-import com.android.volley.RequestQueue;
-import com.android.volley.Response;
-import com.android.volley.VolleyError;
-import com.android.volley.toolbox.JsonObjectRequest;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.core.app.ActivityOptionsCompat;
+import androidx.core.util.Pair;
+import androidx.fragment.app.Fragment;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 
-import org.json.JSONObject;
+import com.android.volley.RequestQueue;
+import com.android.volley.toolbox.JsonObjectRequest;
 
 import java.util.List;
 
@@ -65,46 +63,39 @@ public class CastFragment extends Fragment implements View.OnClickListener, Cast
     TextView card_holder;
     @BindView(R.id.breathingProgressFragment)
     BreathingProgress breathingProgress;
-    private String cast_json;
+    @BindView(R.id.detail_fragment_views_layout)
+    RelativeLayout relativeLayout;
+    private String jsonCast;
     private String movieId, movieTitle;
     private GotCrewListener gotCrewListener;
-    private String api_key = BuildConfig.API_KEY;
 
-
-    public static CastFragment newInstance(String movie_Id, String movie_Title) {
+    public static CastFragment newInstance(String movieId, String movieTitle) {
         CastFragment fragment = new CastFragment();
         Bundle args = new Bundle();
-        args.putString("movie_id", movie_Id);
-        args.putString("movie_title", movie_Title);
+        args.putString("movie_id", movieId);
+        args.putString("movie_title", movieTitle);
         fragment.setArguments(args);
         return fragment;
     }
 
     @Nullable
     @Override
-    public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
+    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
 
         View view = inflater.inflate(R.layout.cast_fragment, container, false);
         ButterKnife.bind(this, view);
 
         cast_recycler.setLayoutManager(new LinearLayoutManager(getActivity()));
         cast_recycler.setNestedScrollingEnabled(false);
-
         cast_recycler.setVisibility(View.INVISIBLE);
-
         more.setOnClickListener(this);
-
 
         return view;
     }
 
-
-    public void setGotCrewListener(GotCrewListener gotCrewListener){
-
+    public void setGotCrewListener(GotCrewListener gotCrewListener) {
         this.gotCrewListener = gotCrewListener;
-
     }
-
 
     @Override
     public void onCreate(@Nullable Bundle savedInstanceState) {
@@ -113,122 +104,90 @@ public class CastFragment extends Fragment implements View.OnClickListener, Cast
         Bundle savedBundle = getArguments();
 
         if (savedBundle != null) {
-
             movieId = savedBundle.getString("movie_id");
             movieTitle = savedBundle.getString("movie_title");
-
         }
 
-
-        if (movieId != null)
-            getCastFromNetwork(movieId);
-
+        if (movieId != null) getCastFromNetwork(movieId);
     }
-
 
     public void getCastFromNetwork(String movieId) {
 
+        String TMBD_API_KEY = BuildConfig.TMDB_API_KEY;
+        final String BASE_MOVIE_CAST_DETAILS = "http://api.themoviedb.org/3/movie/" + movieId + "/casts?api_key=" + TMBD_API_KEY;
+        JsonObjectRequest jsonObjectRequestForMovieCastDetails = new JsonObjectRequest(BASE_MOVIE_CAST_DETAILS, null,
+                response -> {
+                    jsonCast = response.toString();
+                    parseCastOutput(response.toString());
 
-        final String BASE_MOVIE_CAST_DETAILS = new String("http://api.themoviedb.org/3/movie/" + movieId + "/casts?api_key="+api_key);
-        JsonObjectRequest jsonObjectRequestForMovieCastDetails = new JsonObjectRequest(Request.Method.GET, BASE_MOVIE_CAST_DETAILS, null,
-                new Response.Listener<JSONObject>() {
-                    @Override
-                    public void onResponse(JSONObject response) {
+                    if (gotCrewListener != null) gotCrewListener.gotCrew(response.toString());
 
-                        cast_json = response.toString();
+                }, error -> {
 
-                        cast_parseOutput(response.toString());
-
-                        if (gotCrewListener!=null)
-                          gotCrewListener.gotCrew(response.toString());
-
-                    }
-                }, new Response.ErrorListener() {
-            @Override
-            public void onErrorResponse(VolleyError error) {
-
-                Log.e("webi", "Volley Error: " + error.getCause());
-
-                breathingProgress.setVisibility(View.GONE);
-
-            }
-        }
+                    Log.e("webi", "Volley Error: " + error.getCause());
+                    breathingProgress.setVisibility(View.GONE);
+                }
         );
-
 
         TmdbVolleySingleton volleySingleton = TmdbVolleySingleton.getInstance();
         RequestQueue requestQueue = volleySingleton.getRequestQueue();
         requestQueue.add(jsonObjectRequestForMovieCastDetails);
     }
 
+    private void parseCastOutput(String castResult) {
 
-
-    private void cast_parseOutput(String cast_result) {
-
-        MovieDetailsActivityParseWork par = new MovieDetailsActivityParseWork(getActivity(), cast_result);
-        List<CastDetailsData> cast_list = par.parse_cast();
-        CastAdapter cast_adapter = new CastAdapter(getActivity(), cast_list, true);
+        MovieDetailsActivityParseWork par = new MovieDetailsActivityParseWork(getActivity(), castResult);
+        List<CastDetailsData> castList = par.parse_cast();
+        CastAdapter cast_adapter = new CastAdapter(getActivity(), castList, true);
         cast_adapter.setClickListener(this);
         cast_recycler.setAdapter(cast_adapter);
-        if (cast_list.size() > 4)
+
+        if (castList.size() > 4) {
             more.setVisibility(View.VISIBLE);
-        else if (cast_list.size() == 0) {
+        } else if (castList.size() == 0) {
             more.setVisibility(View.INVISIBLE);
             card_holder.setVisibility(View.INVISIBLE);
-        } else
+        } else {
             more.setVisibility(View.INVISIBLE);
+        }
 
         breathingProgress.setVisibility(View.GONE);
         cast_recycler.setVisibility(View.VISIBLE);
-
-
+        relativeLayout.setMinimumHeight(0);
     }
 
     @Override
     public void onClick(View view) {
 
         if (view.getId() == R.id.more) {
-
             Log.d("webi", "" + movieTitle);
-
-            if (cast_json != null && movieTitle != null) {
-
-
+            if (jsonCast != null && movieTitle != null) {
                 Intent intent = new Intent(getActivity(), FullCastActivity.class);
-                intent.putExtra("cast_json", cast_json);
+                intent.putExtra("cast_json", jsonCast);
                 intent.putExtra("toolbar_title", movieTitle);
                 startActivity(intent);
-
             }
         }
-
-
     }
 
     @Override
     public void itemClicked(CastDetailsData setterGetter, int position, View view) {
         Intent intent = new Intent(getActivity(), CharacterDetailsActivity.class);
-        intent.putExtra("id", setterGetter.getCast_id());
+        intent.putExtra("id", setterGetter.getCastId());
 
         if (Build.VERSION.SDK_INT > Build.VERSION_CODES.KITKAT) {
-
             Pair<View, String> p1 = Pair.create(view.findViewById(R.id.cast_poster), "profile");
             Pair<View, String> p2 = Pair.create(view.findViewById(R.id.cast_name), "name");
 
             ActivityOptionsCompat options = ActivityOptionsCompat.
                     makeSceneTransitionAnimation(getActivity(), p1, p2);
             startActivity(intent, options.toBundle());
-
         } else {
             startActivity(intent);
         }
-
     }
 
-
-
-    public interface GotCrewListener{
+    public interface GotCrewListener {
         void gotCrew(String crewData);
     }
-
 }
