@@ -1,25 +1,26 @@
 package tech.salroid.filmy.fragment;
 
-import android.app.AlertDialog;
 import android.content.Context;
-import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.res.Configuration;
 import android.database.Cursor;
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.LoaderManager;
-import android.support.v4.content.CursorLoader;
-import android.support.v4.content.Loader;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.StaggeredGridLayoutManager;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ArrayAdapter;
 import android.widget.ImageView;
 import android.widget.LinearLayout;
-import android.widget.TextView;
+
+import androidx.annotation.NonNull;
+import androidx.fragment.app.Fragment;
+import androidx.loader.app.LoaderManager;
+import androidx.loader.content.CursorLoader;
+import androidx.loader.content.Loader;
+import androidx.recyclerview.widget.RecyclerView;
+import androidx.recyclerview.widget.StaggeredGridLayoutManager;
+
+import com.google.android.material.dialog.MaterialAlertDialogBuilder;
 
 import butterknife.BindView;
 import butterknife.ButterKnife;
@@ -27,6 +28,7 @@ import tech.salroid.filmy.R;
 import tech.salroid.filmy.activities.MovieDetailsActivity;
 import tech.salroid.filmy.custom_adapter.SavedMoviesAdapter;
 import tech.salroid.filmy.database.FilmContract;
+import tech.salroid.filmy.utility.Constants;
 
 /*
  * Filmy Application for Android
@@ -72,14 +74,14 @@ public class SavedMovies extends Fragment implements LoaderManager.LoaderCallbac
             FilmContract.SaveEntry.SAVE_CERTIFICATION,
             FilmContract.SaveEntry.SAVE_RUNTIME,
             FilmContract.SaveEntry.SAVE_POSTER_LINK,
+            FilmContract.SaveEntry.SAVE_FLAG
     };
 
     private SavedMoviesAdapter mainActivityAdapter;
 
     @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+    public View onCreateView(@NonNull LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-
 
         View view = inflater.inflate(R.layout.fragment_saved_movies, container, false);
         ButterKnife.bind(this, view);
@@ -128,7 +130,10 @@ public class SavedMovies extends Fragment implements LoaderManager.LoaderCallbac
     @Override
     public Loader<Cursor> onCreateLoader(int id, Bundle args) {
 
-        return new CursorLoader(getActivity(), FilmContract.SaveEntry.CONTENT_URI, GET_SAVE_COLUMNS, null, null, "_ID DESC");
+        String selection = FilmContract.SaveEntry.TABLE_NAME+"."+ FilmContract.SaveEntry.SAVE_FLAG + "= ?";
+        String[] selectionArgs = {String.valueOf(Constants.FLAG_OFFLINE)};
+
+        return new CursorLoader(getActivity(), FilmContract.SaveEntry.CONTENT_URI, GET_SAVE_COLUMNS, selection, selectionArgs, "_ID DESC");
     }
 
     @Override
@@ -168,34 +173,30 @@ public class SavedMovies extends Fragment implements LoaderManager.LoaderCallbac
     public void itemLongClicked(final Cursor mycursor, final int position) {
 
 
-        AlertDialog.Builder adb = new AlertDialog.Builder(getActivity());
+       MaterialAlertDialogBuilder adb = new MaterialAlertDialogBuilder(getActivity());
         final ArrayAdapter<String> arrayAdapter = new ArrayAdapter<>(getActivity(), android.R.layout.simple_list_item_1);
-
         arrayAdapter.add("Remove");
-
-
         final Context context = getActivity();
 
-        adb.setAdapter(arrayAdapter, new DialogInterface.OnClickListener() {
-            @Override
-            public void onClick(DialogInterface dialogInterface, int i) {
+        adb.setAdapter(arrayAdapter, (dialogInterface, i) -> {
 
-                final String deleteSelection = FilmContract.SaveEntry.TABLE_NAME + "." + FilmContract.SaveEntry.SAVE_ID + " = ? ";
+            final String deleteSelection = FilmContract.SaveEntry.TABLE_NAME + "." + FilmContract.SaveEntry.SAVE_ID + " = ? AND "+
+            FilmContract.SaveEntry.TABLE_NAME + "." + FilmContract.SaveEntry.SAVE_FLAG + " = ? ";
 
+            int flag_index = mycursor.getColumnIndex(FilmContract.SaveEntry.SAVE_FLAG);
+            int flag = mycursor.getInt(flag_index);
 
-                final String[] deletionArgs = {mycursor.getString(mycursor.getColumnIndex(FilmContract.SaveEntry.SAVE_ID))};
+            final String[] deletionArgs = {mycursor.getString(mycursor.getColumnIndex(FilmContract.SaveEntry.SAVE_ID)), String.valueOf(flag)};
 
-                long deletion_id = context.getContentResolver().delete(FilmContract.SaveEntry.CONTENT_URI, deleteSelection, deletionArgs);
+            long deletion_id = context.getContentResolver().delete(FilmContract.SaveEntry.CONTENT_URI, deleteSelection, deletionArgs);
 
-                if (deletion_id != -1) {
+            if (deletion_id != -1) {
 
-                    mainActivityAdapter.notifyItemRemoved(position);
+                mainActivityAdapter.notifyItemRemoved(position);
 
-                    if (mainActivityAdapter.getItemCount() == 1)
-                        my_saved_movies_recycler.setVisibility(View.GONE);
+                if (mainActivityAdapter.getItemCount() == 1)
+                    my_saved_movies_recycler.setVisibility(View.GONE);
 
-
-                }
             }
         });
 
