@@ -1,10 +1,12 @@
 package tech.salroid.filmy.fragment.intro_fragments;
 
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import androidx.annotation.NonNull;
+import androidx.fragment.app.Fragment;
 
 import tech.salroid.filmy.R;
 
@@ -27,9 +29,8 @@ import tech.salroid.filmy.R;
 
 public class IntroFragmentB extends Fragment {
 
-
     @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+    public View onCreateView(@NonNull LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         return inflater.inflate(R.layout.fragment_intro_fragment_b, container, false);
     }
