@@ -5,7 +5,6 @@ import android.content.SharedPreferences;
 import android.util.Log;
 
 import com.android.volley.AuthFailureError;
-import com.android.volley.Request;
 import com.android.volley.RequestQueue;
 import com.android.volley.Response;
 import com.android.volley.VolleyError;
@@ -41,7 +40,7 @@ import tech.salroid.filmy.network_stuff.TmdbVolleySingleton;
  */
 public class UnMarkingWatchList {
 
-    private String api_key = BuildConfig.API_KEY;
+    private String api_key = BuildConfig.TMDB_API_KEY;
     private String SESSION_PREF = "SESSION_PREFERENCE";
     private TmdbVolleySingleton tmdbVolleySingleton = TmdbVolleySingleton.getInstance();
     private RequestQueue tmdbrequestQueue = tmdbVolleySingleton.getRequestQueue();
@@ -64,7 +63,7 @@ public class UnMarkingWatchList {
 
         String PROFILE_URI = "https://api.themoviedb.org/3/account?api_key="+api_key+"&session_id=" + session_id;
 
-        JsonObjectRequest jsonObjectRequest = new JsonObjectRequest(Request.Method.GET, PROFILE_URI, null,
+        JsonObjectRequest jsonObjectRequest = new JsonObjectRequest(PROFILE_URI, null,
                 new Response.Listener<JSONObject>() {
                     @Override
                     public void onResponse(JSONObject response) {
@@ -124,7 +123,7 @@ public class UnMarkingWatchList {
         final String mRequestBody = jsonBody.toString();
 
 
-        JsonObjectRequest jsonObjectRequest = new JsonObjectRequest(Request.Method.POST, query, null,
+        JsonObjectRequest jsonObjectRequest = new JsonObjectRequest(query, null,
                 new Response.Listener<JSONObject>() {
                     @Override
                     public void onResponse(JSONObject response) {
